+++ smerc_js_sdk/index.mjs
@@ -1,0 +1,1 @@
+export { SMERCAPIError, SMERCClient } from './client.mjs';